requirements.txt
aiofiles
aioredis
aiosqlite
alembic
amqp
anyio
async-generator
async-timeout
asyncpg
attrs
autoflake
Babel
bcrypt
billiard
celery
certifi
cffi
click
click-didyoumean
click-plugins
click-repl
cryptography
dnspython
email-validator
exceptiongroup
fastapi
fastapi-admin
fastapi-cache2
fastapi-users
fastapi-users-db-sqlalchemy
fastapi_amis_admin
fastapi_user_auth
greenlet
h11
httpcore
httptools
httpx
idna
importlib-metadata
importlib-resources
iniconfig
iso8601
itsdangerous
Jinja2
kombu
makefun
Mako
MarkupSafe
orjson
outcome
packaging
passlib
pendulum
pluggy
prompt-toolkit
psycopg2-binary
pycparser
pydantic
pyflakes
PyJWT
pypika-tortoise
pytest
pytest-asyncio
python-dateutil
python-dotenv
python-multipart
pytz
pytzdata
PyYAML
redis
rfc3986
six
sniffio
sortedcontainers
sqladmin
SQLAlchemy
sqlalchemy2-stubs
sqlalchemy_database
sqlmodel
sqlmodelx
starlette
tomli
tortoise-orm
typing_extensions
ujson
uvicorn
uvloop
vine
watchfiles
wcwidth
websockets
WTForms
zipp